Create new class Match4PCSBase - Match4PCSImpl and MatchSuper4PCSImpl inherits Match3PCSBase - Shared fields have been moved to Match4PCSBase
