+++ CMakeLists.txt
@@ -136,6 +136,7 @@ set(Super4PCS_SRC
 )
 set(Super4PCS_INCLUDE
     ${SRC_DIR}/sampling.h
+    ${SRC_DIR}/match4pcsBase.h
     ${SRC_DIR}/shared4pcs.h
     ${SRC_DIR}/pairExtraction.h
     ${SRC_DIR}/4pcs.h
