+++ src/super4pcs/4pcs.cc
@@ -43,6 +43,7 @@
 
 
 #include "4pcs.h"
+#include "match4pcsBase.h"
 #include "sampling.h"
 #include "accelerators/utils.h"
 
@@ -297,37 +298,13 @@ float distSegmentToSegment(const cv::Point3f& p1, const cv::Point3f& p2,
 
 }  // namespace
 
-// Holds a base from P. The base contains 4 points (indices) from the set P.
-struct Quadrilateral {
-  int vertices[4];
-  Quadrilateral(int vertex0, int vertex1, int vertex2, int vertex3) {
-    vertices[0] = vertex0;
-    vertices[1] = vertex1;
-    vertices[2] = vertex2;
-    vertices[3] = vertex3;
-//    cout << vertex0 << " "
-//         << vertex1 << " "
-//         << vertex2 << " "
-//         << vertex3 << " "
-//         << (float(vertex0) + float(vertex1) + float(vertex2) + float(vertex3))/4.f
-//         << endl;
-  }
-};
-
-typedef vector<pair<int, int>> PairsVector;
-
-class Match4PCSImpl {
+class Match4PCSImpl : public Super4PCS::Match4PCSBase {
  public:
+  using Base = Super4PCS::Match4PCSBase;
   explicit Match4PCSImpl(const Match4PCSOptions& options)
-      : number_of_trials_(0),
-        max_base_diameter_(-1),
+        : Base (options),
         ann_tree_(0),
-        data_points_(),
-        P_mean_distance_(1.0),
-        best_LCP_(0.0F),
-        options_(options) {
-    base_3D_.resize(4);
-  }
+        data_points_() {}
 
   ~Match4PCSImpl() {
     // Release the ANN data structure and points.
@@ -362,57 +339,10 @@ class Match4PCSImpl {
 
   // Internal data members.
 
-  // Number of trials. Every trial picks random base from P.
-  int number_of_trials_;
-  // Maximum base diameter. It is computed automatically from the diameter of
-  // P and the estimated overlap and used to limit the distance between the
-  // points in the base in P so that the probability to have all points in
-  // the base as inliers is increased.
-  float max_base_diameter_;
-  // The diameter of P.
-  float P_diameter_;
   // ANN structure allows to query arbitrary point for range searching.
   ANNkd_tree* ann_tree_;
   // Holds the ANN data points.
   ANNpointArray data_points_;
-  // Mean distance between points and their nearest neighbor in the set P.
-  // Used to normalize the "delta" which is given in terms of this distance.
-  float P_mean_distance_;
-  // The centroid about which we rotate a congruent set in Q to match the base
-  // in P. It is used temporarily and makes the transformations more robust to
-  // noise. At the end, the direct transformation applied as a 4x4 matrix on
-  // every points in Q is computed and returned.
-  cv::Point3f centroid_;
-  // The translation vector by which we move Q to match P.
-  cv::Point3f translate_;
-  // The rotation matrix by which we rotate Q to match P.
-  cv::Mat rotate_;
-  // The points in the base (indices to P). It is being updated in every
-  // RANSAC iteration.
-  int base_[4];
-  // The current congruent 4 points from Q. Every RANSAC iteration the
-  // algorithm examines a set of such congruent 4-points from Q and retains
-  // the best from them (the one that realizes the best LCP).
-  int current_congruent_[4];
-  // Sampled P (3D coordinates).
-  std::vector<Point3D> sampled_P_3D_;
-  // Sampled Q (3D coordinates).
-  std::vector<Point3D> sampled_Q_3D_;
-  // The 3D points of the base.
-  std::vector<Point3D> base_3D_;
-  // The copy of the input Q. We transform Q to match P and returned the
-  // transformed version.
-  std::vector<Point3D> Q_copy_;
-  // The centroid of P.
-  cv::Point3f centroid_P_;
-  // The centroid of Q.
-  cv::Point3f centroid_Q_;
-  // The best LCP (Largest Common Point) fraction so far.
-  float best_LCP_;
-  // Current trial.
-  int current_trial_;
-  // Parameters.
-  Match4PCSOptions options_;
 
   // Private member functions.
 
@@ -436,7 +366,7 @@ class Match4PCSImpl {
   // respect to distance and normals, up to the given tolerance.
   void BruteForcePairs(double pair_distance, double pair_normals_angle,
                        double pair_distance_epsilon, int base_point1,
-                       int base_point2, PairsVector* pairs);
+                       int base_point2, Base::PairsVector* pairs);
 
   // For each randomly picked base, verifies the computed transformation by
   // computing the number of points that this transformation brings near points
@@ -484,7 +414,7 @@ class Match4PCSImpl {
                                    double distance_threshold2,
                                    const PairsVector& P_pairs,
                                    const PairsVector& Q_pairs,
-                                   std::vector<Quadrilateral>* quadrilaterals);
+                                   std::vector<Super4PCS::Quadrilateral>* quadrilaterals);
 
   // Initializes the data structures and needed values before the match
   // computation.
@@ -506,7 +436,7 @@ bool Match4PCSImpl::FindCongruentQuadrilaterals(
     double invariant1, double invariant2, double distance_threshold1,
     double distance_threshold2, const std::vector<std::pair<int, int>>& P_pairs,
     const std::vector<std::pair<int, int>>& Q_pairs,
-    std::vector<Quadrilateral>* quadrilaterals) {
+    std::vector<Super4PCS::Quadrilateral>* quadrilaterals) {
   if (quadrilaterals == NULL) return false;
 
   int number_of_points = 2 * P_pairs.size();
@@ -556,7 +486,7 @@ bool Match4PCSImpl::FindCongruentQuadrilaterals(
         const Point3D& pp2 = sampled_Q_3D_[P_pairs[id].second];         
                     
         quadrilaterals->push_back(
-            Quadrilateral(P_pairs[id].first, P_pairs[id].second,
+            Super4PCS::Quadrilateral(P_pairs[id].first, P_pairs[id].second,
                           Q_pairs[i].first, Q_pairs[i].second));
       } else
         break;
@@ -578,7 +508,7 @@ bool Match4PCSImpl::FindCongruentQuadrilaterals(
         const Point3D& pp2 = sampled_Q_3D_[P_pairs[id].second];   
         
         quadrilaterals->push_back(
-            Quadrilateral(P_pairs[id].first, P_pairs[id].second,
+            Super4PCS::Quadrilateral(P_pairs[id].first, P_pairs[id].second,
                           Q_pairs[i].first, Q_pairs[i].second));
       } else
         break;
@@ -948,7 +878,7 @@ bool Match4PCSImpl::TryOneBase() {
   double distance2 = PointsDistance(base_3D_[2], base_3D_[3]);
 
   vector<pair<int, int>> pairs1, pairs2;
-  vector<Quadrilateral> congruent_quads;
+  vector<Super4PCS::Quadrilateral> congruent_quads;
 
   // Compute normal angles.
   double normal_angle1 = (base_3D_[0].normal() - base_3D_[1].normal()).norm();
